requirements.txt
telepot==4.1
python-decouple==3.0
pyserial==2.7
